Extract HmacKeyGen from Hmac
