+++ crypto/js/src/main/scala/org/http4s/crypto/CryptoPlatform.scala
@@ -24,5 +24,6 @@ private[crypto] trait CryptoCompanionPlatform {
     new UnsealedCrypto[F] {
       override def hash: Hash[F] = Hash[F]
       override def hmac: Hmac[F] = Hmac[F]
+      override def hmacKeyGen: HmacKeyGen[F] = HmacKeyGen[F]
     }
 }
